Add utility function `constantly`.
Works like Clojure's constantly; that is, given a value, it creates a
function which accepts any number of arguments and returns that value
regardless of what's passed in. Potentially useful for code which
takes callbacks when all you want to pass in is a value; in other words,
for separating concerns in the eventual entity-component system.