Store map entities on a tile as list.
Should now be able to have more than one entity on a tile, which is
necessary to have items and combat-dogpiling (which probably won't
actually come up in this roguelike, I think, but who knows) and tiles as
entities.

In addition, passability lookups in the actor system loop should be a
bit more efficient, as it now checks only the tile it's actually
interested in, rather than all tiles in existence.