Rework input-actions map.
In the new system, I'm thinking I'll map input characters to action
tuples which will be lined up in the player-character's action queue. I
may also want to allow mapping to functions, in which case the function
should take no arguments (probably; I may change this so that it takes
the entity manager or work out a similar system) and return an
action-tuple to be pushed onto the queue.

I'm not sure if this will be how I ultimately end up doing it, but it's
a possibility. I think it'll work out, but I honestly have no idea in
practice. We'll see.