Add color-transforming functions.
to_grayscale converts any color to grayscale.

color_to_parchment_tone converts a given color to a parchment tone color
for use in the level map (which will be used to show the player the
level as they remember each part; should be able to show tiles and such,
at least. I plan to add some noise to this color. It should look pretty
neat once it's finished... but that's entirely tangential to this
commit.