Add lighting functions.
distance_squared is a helper function which does exactly what it says.

lighting_const just returns i. It will probably be removed at some
point, or perhaps used for calculating ambient lighting in daytime maps,
but honestly will probably not be used at all.

lighting_linear calculates the linear falloff for the tile, which is the
inverse of the distance.

lighting_quadratic calculates the quadratic falloff for the tile, which
is the inverse of the distance squared.

lighting_quadratic_spec calculates the quadratic falloff for the tile,
then multipliers it by two, takes the inverse, and multipliers the
quadratic lighting value for the tile by this inverse and returns it.