Lookup AI functions in module instead of via dict.
Since I'm doing it with other areas, why not AI? I did want a simpler
system, after all. I moved the AI functions to eldestrl.ai.types and
changed the AI system to get the AI type from eldestrl.ai.types,
throwing an error if it the AI component's type attribute names one of
the module's private functions (one that starts with '_'). This should
work well enough, I think.

For now, it raises a generic Exception. I'll want to change this to some
other type at some point, but for now it's fine, as I don't expect any
errors to be raised in the near future. Or ever, really. But it's good
practice to have a specific error type so I can catch it if necessary.