Add mapgen-related utility functions.
outside() and tree_outside() calculate if a rectangle is inside or
outside using a tilemap. count_in_rect() and rect_area() are utility
functions which do exactly what they sound like they'd do; rect_area()
returns the rect's area, and count_in_rect() takes a rect and a map of
truthy values and returns the number of tiles inside the rect's bounds.