Add a test script version of the fog system.
I think this may be the automaton I'll go with, but I'd like to refine
it a bit if I can before I use it in the fog system. It's presently a
bit CPU-intensive and also doesn't spread the fog as nicely as I'd like,
which is a problem, but it otherwise seems to work, which is definitely
a good spot. Hooray.