Implement the fog system at long last.
It doesn't look a hell of a lot like fog yet, unfortunately. I'm gonna
have to work on that. Maybe add an SDL renderer on top for that. But it
works! Sort of. I might want to work on getting it to transition more
smoothly, though. That's a definite priority. That's probably the core
problem rather than how it renders on the map, though fixing the
rendering would also help enormously.

So, yeah. Fog!