Update fog system.
It now uses the algorithm used in the test script, which looks much
nicer than the previous system. I've tweaked some functions involved and
some values so that it looks a bit nicer, but the real gains are going
to come from implementing an SDL renderer hook, if that's even possible
using TDL. We'll see.

Mapgen next? AI? Combat? Buyers? One of these. Probably mapgen, because
mapgen is almost nonexistent right now.